+++ src/time.rs
@@ -3,10 +3,10 @@ use std::time::SystemTime;
 pub struct Time {}
 
 impl Time {
-  pub fn now_millis() -> f64 {
+  pub fn now_millis() -> f32 {
     SystemTime::now()
       .duration_since(SystemTime::UNIX_EPOCH)
       .expect("Time went backwards")
-      .as_micros() as f64 / 1000.
+      .as_micros() as f32 / 1000.
   }
 }
